refactor: css 속성 개행, util 파일 명 수정, useOutsideClick 코리 반영
