+++ src/components/shared/Header/Menu.tsx
@@ -1,8 +1,7 @@
 import { useOutsideClick } from "@/hooks/useOutsideClick";
-import { checkContains } from "@/utils/domUtils";
-import { camelToKebab } from "@/utils/stringUtils";
+import { camelToKebab } from "@/utils/string";
 import type React from "react";
-import { useMemo, useRef, useState } from "react";
+import { useMemo, useState } from "react";
 import { buttonStyle } from "./Header.css";
 
 type MenuProps = {
@@ -11,24 +10,14 @@ type MenuProps = {
   renderList: React.ReactNode;
 };
 
-export type MenuRefType = {
-  setState: React.Dispatch<React.SetStateAction<boolean>>;
-  ref: React.RefObject<HTMLDivElement>;
-};
-
 const Menu = ({ label, renderTriggerIcon, renderList }: MenuProps) => {
   const [showMenu, setShowMenu] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
   const handleClick = () => setShowMenu(!showMenu);
-  const handleOutsideClick = ({ target }: MouseEvent) => {
-    const [check] = checkContains(target!, containerRef);
-    if (!check) setShowMenu(false);
-  };
   const id = useMemo(() => camelToKebab(`${label}Toggle`), [label]);
-  useOutsideClick(handleOutsideClick);
+  const ref = useOutsideClick(setShowMenu);
 
   return (
-    <div ref={containerRef} aria-label="menu container">
+    <div ref={ref} aria-label="menu container">
       <button
         className={buttonStyle}
         id={id}
