+++ src/utils/string.ts
